+++ CHANGELOG.md
@@ -230,6 +230,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 ### Removed
 
 - Delete the following functions and methods
+  - ContentsRef::from\_bytes()
   - ContentsRef::from\_mut\_bytes()
 - Delete the following macros
   - constructed\_ber!()